feat: Back-end finished, initial front-end preparations, and some progress on the login part.

+++ back-end/src/modules/contacts/contacts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateContactDto } from './dto/create-contact.dto';
 import { UpdateContactDto } from './dto/update-contact.dto';
 import { ContactsRepository } from './repositories/contacts.repository';
@@ -19,10 +19,18 @@ export class ContactsService {
   }
 
   async update(id: string, updateContactDto: UpdateContactDto) {
+    const findcontact = await this.contactsRepository.findOne(id);
+    if (!findcontact) {
+      throw new NotFoundException('contact not found');
+    }
     return await this.contactsRepository.update(id, updateContactDto);
   }
 
   async remove(id: string) {
+    const findcontact = await this.contactsRepository.findOne(id);
+    if (!findcontact) {
+      throw new NotFoundException('contact not found');
+    }
     return await this.contactsRepository.delete(id);
   }
 }
